requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiolimiter==1.2.1
aiosignal==1.4.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
audioop-lts==0.2.2
Brotli==1.1.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.3
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.8.0
fastapi==0.121.0
ffmpy==0.6.4
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.10.0
gradio==5.49.1
gradio_client==1.13.3
groovy==0.1.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==1.0.1
idna==3.11
Jinja2==3.1.6
jiter==0.11.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain==1.0.3
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.0.3
langchain-mcp-adapters==0.1.12
langchain-mongodb==0.7.1
langchain-openai==1.0.2
langchain-text-splitters==1.0.0
langchain-voyageai==0.1.3
langgraph==1.0.2
langgraph-checkpoint==2.1.2
langgraph-checkpoint-mongodb==0.2.1
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langgraph-store-mongodb==0.1.0
langsmith==0.4.40
lark==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
mcp==1.20.0
mdurl==0.1.2
multidict==6.7.0
mypy_extensions==1.1.0
numpy==2.3.4
openai==2.7.0
orjson==3.11.4
ormsgpack==1.11.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
propcache==0.4.1
pycparser==2.23
pydantic==2.11.10
pydantic-settings==2.11.0
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.15.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.28.0
ruff==0.14.3
safehttpx==0.1.7
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
sse-starlette==3.0.3
starlette==0.49.3
tenacity==9.1.2
tiktoken==0.12.0
tokenizers==0.22.1
tomlkit==0.13.3
tqdm==4.67.1
typer==0.20.0
typer-slim==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
voyageai==0.3.5
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
